requirements.txt
aiohttp==3.9.1
asyncpg==0.29.0
boto3==1.34.0
openai==1.3.3
pandas==2.1.3
python-dotenv==1.0.0
tqdm==4.66.1